+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [5.0.3] - 2023-04-03
+### Added
+- Add Force Scheduling process
+- Add delete schedule (when schedule deleted all task related to this schedule will be deleted from JIRA)
+- Add Schedule swapping between two team members (when schedule swapped all task related to the old member will be assigned to the new member in JIRA)
+
+
 ## [5.0.2] - 2023-03-21
 ### Added
 - Add email in team member data